Enemy movement, code clean up, few "feature" fixes
Many parts of code that were not needed or commented out were removed to
shorten up the length of the file. More work done on the movement of the
enemies, now all but the weird spawning one moves correctly and without
error. error with how the file is read in fixed.